# contracts/v2/openapi-v2.json
{
  "components": {
    "schemas": {
...
      "MapSetViewAction": {
        "additionalProperties": false,
        "properties": {
          "bbox": {
            "$ref": "#/components/schemas/SpatialBoundingBox"
          },
          "type": {
            "const": "map.set_view",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "bbox"
        ],
        "title": "MapSetViewAction",
        "type": "object"
      },
...
      "ResetRequest": {
        "additionalProperties": false,
        "properties": {
          "seed": {
            "anyOf": [
              {
                "maximum": 2147483647.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "task_ref": {
            "$ref": "#/components/schemas/TaskRef"
          }
        },
        "required": [
          "task_ref"
        ],
        "title": "ResetRequest",
        "type": "object"
      },
      "ResetResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EpisodeResultData"
          },
          "meta": {
            "$ref": "#/components/schemas/V2ApiMeta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "ResetResponse",
        "type": "object"
      },
...
      "StateData": {
        "additionalProperties": false,
        "properties": {
          "episode_id": {
            "pattern": "^ep2-[a-f0-9]{32}$",
            "title": "Episode Id",
            "type": "string"
          },
          "semantic_state_hash": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Semantic State Hash",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/V2EpisodeState"
          },
          "state_hash": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "State Hash",
            "type": "string"
          }
        },
        "required": [
          "episode_id",
          "state",
          "state_hash",
          "semantic_state_hash"
        ],
        "title": "StateData",
        "type": "object"
      },
      "StateResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/StateData"
          },
          "meta": {
            "$ref": "#/components/schemas/V2ApiMeta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "StateResponse",
        "type": "object"
      },
      "StepRequest": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "discriminator": {
              "mapping": {
                "answer.abstain": "#/components/schemas/AnswerAbstainAction",
                "answer.request_human_review": "#/components/schemas/AnswerHumanReviewAction",
                "answer.submit": "#/components/schemas/AnswerSubmitAction",
                "map.layer.set_opacity": "#/components/schemas/MapLayerOpacityAction",
                "map.layer.set_visibility": "#/components/schemas/MapLayerVisibilityAction",
                "map.pan": "#/components/schemas/MapPanAction",
                "map.set_view": "#/components/schemas/MapSetViewAction",
                "map.time.set_range": "#/components/schemas/MapTimeRangeAction",
                "map.zoom": "#/components/schemas/MapZoomAction",
                "memory.bookmark_aoi": "#/components/schemas/MemoryBookmarkAoiAction",
                "memory.save_evidence": "#/components/schemas/MemorySaveEvidenceAction",
                "tool.invoke": "#/components/schemas/ToolInvokeAction"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/MapSetViewAction"
              },
              {
                "$ref": "#/components/schemas/MapPanAction"
              },
              {
                "$ref": "#/components/schemas/MapZoomAction"
              },
              {
                "$ref": "#/components/schemas/MapLayerVisibilityAction"
              },
              {
                "$ref": "#/components/schemas/MapLayerOpacityAction"
              },
              {
                "$ref": "#/components/schemas/MapTimeRangeAction"
              },
              {
                "$ref": "#/components/schemas/MemorySaveEvidenceAction"
              },
              {
                "$ref": "#/components/schemas/MemoryBookmarkAoiAction"
              },
              {
                "$ref": "#/components/schemas/ToolInvokeAction"
              },
              {
                "$ref": "#/components/schemas/AnswerSubmitAction"
              },
              {
                "$ref": "#/components/schemas/AnswerAbstainAction"
              },
              {
                "$ref": "#/components/schemas/AnswerHumanReviewAction"
              }
            ],
            "title": "Action"
          },
          "client_action_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
            "title": "Client Action Id",
            "type": "string"
          },
          "expected_state_version": {
            "minimum": 0.0,
            "title": "Expected State Version",
            "type": "integer"
          }
        },
        "required": [
          "client_action_id",
          "expected_state_version",
          "action"
        ],
        "title": "StepRequest",
        "type": "object"
      },
      "StepResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EpisodeResultData"
          },
          "meta": {
            "$ref": "#/components/schemas/V2ApiMeta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "StepResponse",
        "type": "object"
      },
...
      "TraceResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TraceData"
          },
          "meta": {
            "$ref": "#/components/schemas/V2ApiMeta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "TraceResponse",
        "type": "object"
      },
...
      "V2ApiMeta": {
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "const": "v2",
            "title": "Api Version",
            "type": "string"
          },
          "request_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
            "title": "Request Id",
            "type": "string"
          },
          "schema": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Schema",
            "type": "string"
          },
          "schema_version": {
            "const": "2.0.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "api_version",
          "schema_version",
          "schema",
          "request_id"
        ],
        "title": "V2ApiMeta",
        "type": "object"
      },
...
      "V2ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/V2ApiError"
          },
          "meta": {
            "$ref": "#/components/schemas/V2ApiMeta"
          }
        },
        "required": [
          "meta",
          "error"
        ],
        "title": "V2ErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Task-driven, evidence-native EO agent environment contract. Schema version 2.0.0.",
    "title": "EO Harness Environment API V2",
    "version": "0.4.0",
    "x-api-version": "v2",
    "x-schema-version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v2/episodes/{episode_id}/state": {
      "get": {
        "operationId": "get_state_v2_episodes__episode_id__state_get",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "pattern": "^ep2-[a-f0-9]{32}$",
              "title": "Episode Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Task policy rejected action"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "State or idempotency conflict"
          },
          "416": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Artifact range is not satisfiable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Internal service error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 service unavailable"
          }
        },
        "summary": "Get State",
        "tags": [
          "V2 Environment"
        ]
      }
    },
    "/v2/episodes/{episode_id}/step": {
      "post": {
        "operationId": "step_v2_episodes__episode_id__step_post",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "pattern": "^ep2-[a-f0-9]{32}$",
              "title": "Episode Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Task policy rejected action"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "State or idempotency conflict"
          },
          "416": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Artifact range is not satisfiable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Internal service error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 service unavailable"
          }
        },
        "summary": "Step",
        "tags": [
          "V2 Environment"
        ]
      }
    },
    "/v2/episodes/{episode_id}/trace": {
      "get": {
        "operationId": "get_trace_v2_episodes__episode_id__trace_get",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "pattern": "^ep2-[a-f0-9]{32}$",
              "title": "Episode Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Task policy rejected action"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "State or idempotency conflict"
          },
          "416": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Artifact range is not satisfiable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Internal service error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 service unavailable"
          }
        },
        "summary": "Get Trace",
        "tags": [
          "V2 Environment"
        ]
      }
    },
    "/v2/reset": {
      "post": {
        "operationId": "reset_v2_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Task policy rejected action"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "State or idempotency conflict"
          },
          "416": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Artifact range is not satisfiable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "Internal service error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorResponse"
                }
              }
            },
            "description": "V2 service unavailable"
          }
        },
        "summary": "Reset",
        "tags": [
          "V2 Environment"
        ]
      }
    },
...
  }
}